Reorganize main and prostitution branches for early feedback (#244)
Goal: filter users out early so they don't waste time.

main branch:
- before: main1 > main2 > main3 > main4 > main5 > domestic
- after: main1 > main2 > main5 > main3 > main4 > domestic

prostitution branch:
- before: pro1 > pro2 > pro4 > pro5 > pro6 > pro-eligible
- after: pro1 > pro5 > pro6 > pro2 > pro4 > pro-eligible

I had hoped to add navigation test coverage for this but it proved to be too time consuming because testing-library and jest-dom do not support it out of the box. Will add in a follow up.